requirements.txt
streamlit
streamlit_drawable_canvas
numpy
pandas
Image
plotly
scikit-image
scikit-learn
